💄 검색어 지우기 아이콘 위치 조정 및 선택된 재료컴포넌트 분리
